+++ Productivity_Timer/Productivity Timer/Activities/Extensions/A.TableViewExtension.swift
@@ -27,107 +27,4 @@ extension ActivityTableViewController {
         cell.backgroundColor = blueMoonlight
         return cell
     }
-
-    override func tableView(_ tableView: UITableView, editingStyleForRowAt indexPath: IndexPath) -> UITableViewCell.EditingStyle {
-        return .delete
-    }
-
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        
-        guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { fatalError() }
-        let context: NSManagedObjectContext = appDelegate.persistentContainer.viewContext
-
-        let request = NSFetchRequest<NSFetchRequestResult>(entityName: "Activity")
-        do {
-            let results: NSArray = try context.fetch(request) as NSArray
-            for result in results {
-
-                let activity = result as! Activity
-                if editingStyle == .delete {
-                    activity.deletedDate = Date()
-                    try context.save()
-                }
-            }
-        } catch {
-            print("Fetch failed")
-        }
-
-        if editingStyle == .delete {
-            do {
-                try context.save()
-                ActivitiesObject.arrayOfActivities.remove(at: indexPath.row)
-                tableView.deleteRows(at: [indexPath], with: .fade)
-                tableView.reloadData()
-            } catch {
-                print("Can't save the context")
-            }
-        }
-    }
-
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        return true
-    }
-
-    override func tableView(_ tableView: UITableView, moveRowAt sourceIndexPath: IndexPath, to destinationIndexPath: IndexPath) {
-        let moved = ActivitiesObject.arrayOfActivities.remove(at: sourceIndexPath.row)
-        ActivitiesObject.arrayOfActivities.insert(moved, at: destinationIndexPath.row)
-        tableView.reloadData()
-    }
-
-    override func tableView(_ tableView: UITableView, leadingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-        let done = doneAction(at: indexPath)
-        let favourite = favouriteAction(at: indexPath)
-        return UISwipeActionsConfiguration(actions: [done, favourite])
-    }
-
-    func doneAction(at indexPath: IndexPath) -> UIContextualAction {
-        guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { fatalError() }
-
-        let context: NSManagedObjectContext = appDelegate.persistentContainer.viewContext
-        let request = NSFetchRequest<NSFetchRequestResult>(entityName: "Activity")
-
-        do {
-            let results: NSArray = try context.fetch(request) as NSArray
-            for result in results {
-
-                let activity = result as! Activity
-                activity.deletedDate = Date()
-                try context.save()
-            }
-        } catch {
-            print("Fetch failed")
-        }
-
-        let action = UIContextualAction(style: .destructive, title: "Done") { (action, view, completion) in
-            ActivitiesObject.arrayOfActivities.remove(at: indexPath.row)
-            self.tableView.deleteRows(at: [indexPath], with: .automatic)
-            completion(true)
-            
-            // TODO: coreData access
-            ActivitiesObject.arrayOfActivities.remove(at: indexPath.row)
-            print(ActivitiesObject.arrayOfActivities)
-            self.tableView.reloadData()
-        }
-        action.backgroundColor = .systemGreen
-        action.image = UIImage(systemName: "checkmark.circle")
-        return action
-    }
-
-    func favouriteAction(at indexPath: IndexPath) -> UIContextualAction {
-        let object = ActivitiesObject.arrayOfActivities[indexPath.row]
-        let action = UIContextualAction(style: .normal, title: "Favourite") { (action, view, completion) in
-            object.fav = !object.fav
-            ActivitiesObject.arrayOfActivities[indexPath.row] = object
-            completion(true)
-        }
-
-        action.backgroundColor = object.fav ? .systemPurple : .systemGray
-        action.image = object.fav ? UIImage(systemName: "heart.fill") : UIImage(systemName: "heart")
-        if object.fav {
-            object.fav = true
-        } else {
-            object.fav = false
-        }
-        return action
-    }
 }
